Implemented an image Index update to prevent render failure.
